+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - internal/integration/prosody: add option for enabling bidirectional s2s
   connections
+- roster: add `Set` and `Delete` functions for roster management
 - xmpp: `SetReadDeadline` and `SetWriteDeadline` are now proxied even if the
   underlying connection is not a `net.Conn`
 - xmpp: all sent stanzas are now given randomly generated IDs if no ID was